User Story 3 completed
