+++ components/Layout/index.js
@@ -1,0 +1,5 @@
+import Navigaton from "../Navigation";
+
+export default function Layout() {
+  return <Navigaton />;
+}
