+++ pages/art-pieces/index.js
@@ -1,0 +1,5 @@
+import ArtPieces from "@/components/ArtPieces";
+
+export default function ArtPiecesPage({ apiData }) {
+  return <ArtPieces pieces={apiData}></ArtPieces>;
+}
